+++ packages/webview_flutter/webview_flutter/lib/src/navigation_delegate.dart
@@ -45,6 +45,7 @@ class NavigationDelegate {
         onNavigationRequest,
     void Function(String url)? onPageStarted,
     void Function(String url)? onPageFinished,
+    void Function(int statusCode)? onPageError,
     void Function(int progress)? onProgress,
     void Function(WebResourceError error)? onWebResourceError,
     void Function(UrlChange change)? onUrlChange,
@@ -53,6 +54,7 @@ class NavigationDelegate {
           onNavigationRequest: onNavigationRequest,
           onPageStarted: onPageStarted,
           onPageFinished: onPageFinished,
+          onPageError: onPageError,
           onProgress: onProgress,
           onWebResourceError: onWebResourceError,
           onUrlChange: onUrlChange,
@@ -95,6 +97,7 @@ class NavigationDelegate {
         onNavigationRequest,
     void Function(String url)? onPageStarted,
     void Function(String url)? onPageFinished,
+    void Function(int statusCode)? onPageError,
     void Function(int progress)? onProgress,
     void Function(WebResourceError error)? onWebResourceError,
     void Function(UrlChange change)? onUrlChange,
@@ -103,6 +106,7 @@ class NavigationDelegate {
           onNavigationRequest: onNavigationRequest,
           onPageStarted: onPageStarted,
           onPageFinished: onPageFinished,
+          onPageError: onPageError,
           onProgress: onProgress,
           onWebResourceError: onWebResourceError,
           onUrlChange: onUrlChange,
@@ -116,6 +120,7 @@ class NavigationDelegate {
     this.onNavigationRequest,
     this.onPageStarted,
     this.onPageFinished,
+    this.onPageError,
     this.onProgress,
     this.onWebResourceError,
     void Function(UrlChange change)? onUrlChange,
@@ -128,6 +133,9 @@ class NavigationDelegate {
     }
     if (onPageFinished != null) {
       platform.setOnPageFinished(onPageFinished!);
+    }
+    if (onPageError != null) {
+      platform.setOnPageError(onPageError!);
     }
     if (onProgress != null) {
       platform.setOnProgress(onProgress!);
@@ -160,6 +168,9 @@ class NavigationDelegate {
 
   /// Invoked when a page has finished loading.
   final PageEventCallback? onPageFinished;
+
+  /// Invoked when a page failed loading
+  final PageErrorCallback? onPageError;
 
   /// Invoked when a page is loading to report the progress.
   final ProgressCallback? onProgress;