+++ packages/webview_flutter/webview_flutter/test/legacy/webview_flutter_test.mocks.dart
@@ -94,6 +94,7 @@ class MockWebViewPlatform extends _i1.Mock implements _i4.WebViewPlatform {
           ),
         ),
       ) as _i2.Widget);
+
   @override
   _i9.Future<bool> clearCookies() => (super.noSuchMethod(
         Invocation.method(
@@ -122,6 +123,7 @@ class MockWebViewPlatformController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+
   @override
   _i9.Future<void> loadFlutterAsset(String? key) => (super.noSuchMethod(
         Invocation.method(
@@ -131,6 +133,7 @@ class MockWebViewPlatformController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+
   @override
   _i9.Future<void> loadHtmlString(
     String? html, {
@@ -145,6 +148,7 @@ class MockWebViewPlatformController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+
   @override
   _i9.Future<void> loadUrl(
     String? url,
@@ -161,6 +165,7 @@ class MockWebViewPlatformController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+
   @override
   _i9.Future<void> loadRequest(_i5.WebViewRequest? request) =>
       (super.noSuchMethod(
@@ -171,6 +176,7 @@ class MockWebViewPlatformController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+
   @override
   _i9.Future<void> updateSettings(_i5.WebSettings? setting) =>
       (super.noSuchMethod(
@@ -181,6 +187,7 @@ class MockWebViewPlatformController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+
   @override
   _i9.Future<String?> currentUrl() => (super.noSuchMethod(
         Invocation.method(
@@ -189,6 +196,7 @@ class MockWebViewPlatformController extends _i1.Mock
         ),
         returnValue: _i9.Future<String?>.value(),
       ) as _i9.Future<String?>);
+
   @override
   _i9.Future<bool> canGoBack() => (super.noSuchMethod(
         Invocation.method(
@@ -197,6 +205,7 @@ class MockWebViewPlatformController extends _i1.Mock
         ),
         returnValue: _i9.Future<bool>.value(false),
       ) as _i9.Future<bool>);
+
   @override
   _i9.Future<bool> canGoForward() => (super.noSuchMethod(
         Invocation.method(
@@ -205,6 +214,7 @@ class MockWebViewPlatformController extends _i1.Mock
         ),
         returnValue: _i9.Future<bool>.value(false),
       ) as _i9.Future<bool>);
+
   @override
   _i9.Future<void> goBack() => (super.noSuchMethod(
         Invocation.method(
@@ -214,6 +224,7 @@ class MockWebViewPlatformController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+
   @override
   _i9.Future<void> goForward() => (super.noSuchMethod(
         Invocation.method(
@@ -223,6 +234,7 @@ class MockWebViewPlatformController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+
   @override
   _i9.Future<void> reload() => (super.noSuchMethod(
         Invocation.method(
@@ -232,6 +244,7 @@ class MockWebViewPlatformController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+
   @override
   _i9.Future<void> clearCache() => (super.noSuchMethod(
         Invocation.method(
@@ -241,6 +254,7 @@ class MockWebViewPlatformController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+
   @override
   _i9.Future<String> evaluateJavascript(String? javascript) =>
       (super.noSuchMethod(
@@ -250,6 +264,7 @@ class MockWebViewPlatformController extends _i1.Mock
         ),
         returnValue: _i9.Future<String>.value(''),
       ) as _i9.Future<String>);
+
   @override
   _i9.Future<void> runJavascript(String? javascript) => (super.noSuchMethod(
         Invocation.method(
@@ -259,6 +274,7 @@ class MockWebViewPlatformController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+
   @override
   _i9.Future<String> runJavascriptReturningResult(String? javascript) =>
       (super.noSuchMethod(
@@ -268,6 +284,7 @@ class MockWebViewPlatformController extends _i1.Mock
         ),
         returnValue: _i9.Future<String>.value(''),
       ) as _i9.Future<String>);
+
   @override
   _i9.Future<void> addJavascriptChannels(Set<String>? javascriptChannelNames) =>
       (super.noSuchMethod(
@@ -278,6 +295,7 @@ class MockWebViewPlatformController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+
   @override
   _i9.Future<void> removeJavascriptChannels(
           Set<String>? javascriptChannelNames) =>
@@ -289,6 +307,7 @@ class MockWebViewPlatformController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+
   @override
   _i9.Future<String?> getTitle() => (super.noSuchMethod(
         Invocation.method(
@@ -297,6 +316,7 @@ class MockWebViewPlatformController extends _i1.Mock
         ),
         returnValue: _i9.Future<String?>.value(),
       ) as _i9.Future<String?>);
+
   @override
   _i9.Future<void> scrollTo(
     int? x,
@@ -313,6 +333,7 @@ class MockWebViewPlatformController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+
   @override
   _i9.Future<void> scrollBy(
     int? x,
@@ -329,6 +350,7 @@ class MockWebViewPlatformController extends _i1.Mock
         returnValue: _i9.Future<void>.value(),
         returnValueForMissingStub: _i9.Future<void>.value(),
       ) as _i9.Future<void>);
+
   @override
   _i9.Future<int> getScrollX() => (super.noSuchMethod(
         Invocation.method(
@@ -337,6 +359,7 @@ class MockWebViewPlatformController extends _i1.Mock
         ),
         returnValue: _i9.Future<int>.value(0),
       ) as _i9.Future<int>);
+
   @override
   _i9.Future<int> getScrollY() => (super.noSuchMethod(
         Invocation.method(